requirements.txt
aerich==0.5.0
aiosqlite==0.16.1
asyncpg==0.22.0
click==7.1.2
ddlparse==1.9.0
dictdiffer==0.8.1
fastapi==0.63.0
h11==0.12.0
iso8601==0.1.14
kafka-python==2.0.2
pydantic==1.8
pyparsing==2.4.7
PyPika==0.44.1
pytz==2020.5
redis==3.5.3
rejson==0.5.4
simplejson==3.17.2
six==1.15.0
starlette==0.13.6
tortoise-orm==0.16.21
typing-extensions==3.7.4.3
uvicorn==0.13.4
